Remove unnecessary conditional
This line of code started out life in the RequestContext initializer. Here:
https://github.com/alphagov/bouncer/commit/e0d52a855eb19483ea767511181f31349fce2fa1

This was needed because there were two contexts in which you instantiated
RequestContext:
1) app.rb
2) rules.rb

In the former, you would have a Rack 'env' object. In the latter, you would
have a Rack::Request object. The intent was to be able to handle either
situation.

This hasn't been needed since:
https://github.com/alphagov/bouncer/commit/4de57dbdb17ceadab0b6ed451e57b39f42038fb0

When we began passing in the Rack::Request created in app.rb.