DFID r4d rule retains the original path
Change the rule so that it keeps the capitalisation of the URL requested.
Incidentally, it will now keep anything else about the path that was changed by
canonicalisation.

The immediate need for this is to get a smokey test passing. The smokey test
expects that the capitalisation of the redirect matches the input. The test is
this one:
https://github.com/alphagov/redirector/blob/5481ae103aae8d9ec97c78c2f3df0a3d9c9177d0/data/tests/popular.csv#L912