Add a new rake task to notify Errbit of a deploy
Airbrake includes three rake tasks, one of which does deploy notification:

    https://github.com/airbrake/airbrake/blob/v3.1.15/lib/airbrake/shared_tasks.rb

However, that task only loads configuration from Rails initializers, so doesn't
work for Rack apps unless Airbrake's configuration is already loaded somehow.
Since the other two tasks from the Airbrake gem aren't useful to us here
either, we're no longer including them in the Rakefile (one is Heroku-specific
and the other, to test the setup, is even more Railsy than the deploy one).

Instead, we've added a new rake task to notify Errbit of a deploy, which
is very similar to Airbrake's built-in one but loads our config instead
of looking for a Rails initializer.