Use Addressable to construct URLs for sitemaps
This makes our sitemaps slightly more robust: if mappings were imported
(bypassing canonicalization) with paths containing characters which
URI.parse can't handle, then the sitemap for that site would 500.
With Addressable, a sitemap would be produced, but it may be valid XML
but invalid as a sitemap because of those characters.

We don't currently have any mappings which include []<> and now that
we have background processing for bulk adding it's unlikely that we
will need to use the mappings import for new sites, but we do still
use it to import directgov and businesslink sites from Redirector.