Link to the page not the homepage on TNA for a 404
We currently link to the TNA page for the domain, rather than the TNA page for
the page you requested. The assumption may have been that this page isn't in
the archive, so linking to the archive for the homepage was a better option.

If I am correct about the origins of this behaviour, then it may have been when
we thought our coverage would have been excellent. That is not generally the
case.