Update MCGA site rules
The rules added in 767a7eefe80408b0856c9a7ced0daf1a02bdbfea and
599d934da9e615ad13f70a2c1c02b6660e6ea029 are not being used at the
moment because the site is configured as a global redirect appending
path to www.dft.gov.uk/mca. Because the global type case comes before
the Outcome::Status (which would use the rules otherwise) in app.rb
we aren't getting the behaviour we want.

This change adds a pre-emptive rule to replicate a global-path preserving
redirect, with the intent that the path-preserving config will be removed.