Add a canary route to enable full-stack testing
This enables us to make a request to www.direct.gov.uk/__canary__
which will test that Bouncer can access everything it needs in the
database. It should not be cached, so that we can request it via the
CDN and get an up-to-date response.

No tests here - this is a sketch to record the outcome of my conversation
with Dan C before Christmas :)