Add fallback rules for Land Registry house prices site
These paths contain address and postcode search parameters, so can't be
mapped.

These rules don't preserve the search - they just redirect to the new price
paid dataset search page. Land Registry have confirmed that this is the right
URL to redirect to.

This rule could be a one-liner, but I thought it was clearer as a `case`:

    redirect('http://landregistry.data.gov.uk/app/ppd') if request.path =~ %r{(^/sold-prices)|(^/price-paid-record)}