Update db/structure.sql from Transition
Bouncer and Transition share a database, so we should keep the structure.sql
here up to date when we modify the database schema in Transition. It looks
like we haven't done this for a while, so there are a few changes, although
none which affect tables used by Bouncer.

Generated with `rake db:structure:dump` in Transition and copied over.