Handle invalid hosts gracefully
We're seeing fairly high volumes of errors in Sentry:
https://sentry.io/organizations/govuk/issues/2086648123/events/80d68b1848f24f7b82e08cb839b64dc2/?project=202210

These errors happen when an invalid (malformed) host is passed to
Bouncer, leading to a `Addressable::URI::InvalidURIError` being
raised. We see an error message of the form:
`Invalid character in host: 'foo.gov.uk))'("),,)'`

We can't stop those requests from being made, but we can avoid
logging inactionable errors in Sentry. This change detects the
`Addressable::URI::InvalidURIError` exception and returns a 400
Bad Request error gracefully: https://httpstatuses.com/400

It wasn't straightforward to add a test for this, as any attempt
to `get url` (where `url` is malformed as above) would break at
the RSpec level and not reach our application code. Consequently,
I've had to do some slightly smelly mocking.

The biggest smell is probably the `allow_any_instance_of`, which
is a pragmatic choice that is better than the alternative (moving
the `context` assignment in `Bouncer::App` into its own method and
stubbing that instead). We'd have had to either make calls to the
`context` method throughout the `call` (inefficient) or cache it
with `@context ||=` (could cause bugs). A slightly non-standard
test feels like the least risky approach here.