Rewrite README to favour GOV.UK Docker
Depends on: https://github.com/alphagov/govuk-docker/pull/479

Transition Config has no entry for "dev.bis.gov.uk", so it's unclear
if these instructions ever worked, or indeed what the expected outcome
was supposed to be. For the sake of clarity, I've used a test domain
and provided instructions that don't rely on Transition Config.

Note that there are other features that I've not written instructions
for here. It should be possible to re-uuse the test domain to check they
work with different config for the site in the DB.