Include prometheus middleware in the app
Unless we shoehorn this initializer in somewhere we won't get any metrics.

I'd originally mis-remembered that Bouncer was a Sinatra app, but it isn't,
it's just plain old Rack. Because of this, we need to add the middleware
in a Rack Builder.

We need to use the `prepend` option for instrumentation as the app crashes
if we use the default `alias_method`. Some documentation on this is in the
[prometheus_exporter README](https://github.com/discourse/prometheus_exporter#choosing-the-style-of-method-patching)