Add prometheus metrics to bouncer
We need to do this [differently to Rails, Sinatra](https://github.com/alphagov/govuk_app_config/blob/main/lib/govuk_app_config/govuk_prometheus_exporter.rb)
because we're using a plain Rack middleware stack.

We need to use the `prepend` option for instrumentation as the app crashes
if we use the default `alias_method`. Some documentation on this is in the
[prometheus_exporter README](https://github.com/discourse/prometheus_exporter#choosing-the-style-of-method-patching)

My previous effort worked in terms of extracting the metrics, but seemed
to leak memory horribly. I think I'd just got it very wrong. The workload
stats on integration for this change seem stable after a couple of hours,
so I think this is good to go.