Fix Active Record deprecation
In #461, it was noted that
`ActiveRecord::Base.clear_active_connections!` is deprecated. It looks
as though it's now been fully removed, as we're seeing the following
error in integration (spotted in a degraded Argo CD pod's logs) after
having upgraded to Active Record 7.2.0:

```
NoMethodError: undefined method `clear_active_connections!' for class ActiveRecord::Base
```

We should be able to replace it with
`ActiveRecord::Base.connection_handler.clear_active_connections!`

Similar changes in other (non-GDS) repos:
- https://github.com/bigcommerce/gruf/pull/208
- https://github.com/phusion/passenger/commit/d2a4ada87303565164b8a8802f0c65615ab1e315

That said, for not-obviously-documented reasons, we should be keeping
the Active Record versions in sync between Bouncer and Transition.
Transition remains on Rails 7.1.3.4, and Dependabot hasn't raised an
upgrade PR, so we should probably revert the upgrade here for now. In
any case, it will be useful to have this deprecation fixed in
preparation for an eventual upgrade

See:
- https://github.com/alphagov/bouncer/commit/cae58279aabafd8225f71fa0e90f02a10711e650
- https://github.com/alphagov/bouncer/commit/c6791bdff0ce593df35cfda7d8571ea75dbaa2b6
- https://github.com/alphagov/bouncer/pull/143#discussion_r119315712
- https://github.com/alphagov/transition/commit/e9266720fe92e8b8d9c2771bf4fc30385bba8137#diff-d09ea66f8227784ff4393d88a19836f321c915ae10031d16c93d67e6283ab55fR8
- https://github.com/alphagov/transition/blob/1a22627c6747cbddf27a9927281b442cb62423df/Gemfile#L3

[461]: https://github.com/alphagov/bouncer/issues/461